+++ src/main/java/com/calculator/models/operators/ExponentOperator.java
@@ -4,8 +4,8 @@ import com.calculator.models.PrecedenceLevel;
 
 public class ExponentOperator implements IOperator {
     @Override
-    public int apply(int firstValue, int secondValue) {
-        return (int) Math.pow(firstValue, secondValue); // Use Math.pow for exponentiation
+    public float apply(float firstValue, float secondValue) {
+        return (float) Math.pow(firstValue, secondValue); // Use Math.pow for exponentiation
     }
 
     @Override
