+++ src/main/java/com/calculator/models/operators/IUnaryOperator.java
@@ -1,7 +1,7 @@
 package com.calculator.models.operators;
 
 public interface IUnaryOperator {
-    int apply(int value);
+    float apply(float value);
 
     static String getSymbol() {
         return null;
