+++ src/main/java/com/calculator/models/operators/IncrementOperator.java
@@ -3,7 +3,7 @@ package com.calculator.models.operators;
 public record IncrementOperator(String variable, boolean postIncrement) implements IUnaryOperator {
 
     @Override
-    public int apply(int value) {
+    public float apply(float value) {
         return value + 1;
     }
