+++ src/main/java/com/calculator/services/VariablesManagerService.java
@@ -5,20 +5,20 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class VariablesManagerService {
-    private static final Map<String, Integer> variables = new HashMap<>();
+    private static final Map<String, Float> variables = new HashMap<>();
 
-    public void putVariable(String variable, int value) {
+    public void putVariable(String variable, float value) {
         variables.put(variable, value);
     }
 
-    public int getVariable(String variable) throws InvalidInputException {
+    public float getVariable(String variable) throws InvalidInputException {
         if (variables.containsKey(variable)) {
             return variables.get(variable);
         }
         throw new InvalidInputException(String.format("Variable %s is not found", variable));
     }
 
-    public Map<String, Integer> getVariables() {
+    public Map<String, Float> getVariables() {
         return variables;
     }
 }
